Refactor main function to its own file
