+++ Makefile
@@ -5,13 +5,13 @@ CC=gcc
 CFLAGS=-pthread -Wall -pedantic-errors
 
 all:
-	$(CC) $(CFLAGS) -o dgtpicom dgtpicom.c
+	$(CC) $(CFLAGS) -o dgtpicom dgtpicom.c main.c
 	$(CC) $(CFLAGS) -shared -fPIC -o dgtpicom.so dgtpicom.c
 	
 debug:
-	$(CC) $(CFLAGS) -Ddebug -o dgtpicom dgtpicom.c
+	$(CC) $(CFLAGS) -Ddebug -o dgtpicom dgtpicom.c main.c
 	$(CC) $(CFLAGS) -Ddebug -shared -fPIC -o dgtpicom.so dgtpicom.c
 
 debug2:
-	$(CC) $(CFLAGS) -Ddebug -Ddebug2 -o dgtpicom dgtpicom.c
+	$(CC) $(CFLAGS) -Ddebug -Ddebug2 -o dgtpicom dgtpicom.c main.c
 	$(CC) $(CFLAGS) -Ddebug -Ddebug2 -shared -fPIC -o dgtpicom.so dgtpicom.c
